move db related things into `database` package
